generic: Change operator ids to u8.

+++ src/operator/add.rs
@@ -150,7 +150,7 @@ reverse_1_add!(AD);
 /// The map from [operator::id] to operator information.
 /// The the map results for ADD_CV_OP, ADD_VC_OP, and ADD_VV_OP are set.
 pub(crate) fn set_op_info( op_info_vec : &mut Vec<OpInfo> ) {
-    op_info_vec[ADD_CV_OP] = OpInfo{
+    op_info_vec[ADD_CV_OP as usize] = OpInfo{
         name           : "add_cv".to_string() ,
         forward_0      : float_forward_0_add_cv,
         forward_1      : float_forward_1_add_cv,
@@ -160,7 +160,7 @@ pub(crate) fn set_op_info( op_info_vec : &mut Vec<OpInfo> ) {
         ad_reverse_1   : ad_reverse_1_add_cv,
         arg_var_index  : super::arg_var_index_binary_cv,
      };
-    op_info_vec[ADD_VC_OP] = OpInfo{
+    op_info_vec[ADD_VC_OP as usize] = OpInfo{
         name           : "add_vc".to_string(),
         forward_0      : float_forward_0_add_vc,
         forward_1      : float_forward_1_add_vc,
@@ -170,7 +170,7 @@ pub(crate) fn set_op_info( op_info_vec : &mut Vec<OpInfo> ) {
         ad_reverse_1   : ad_reverse_1_add_vc,
         arg_var_index  : super::arg_var_index_binary_vc,
     };
-    op_info_vec[ADD_VV_OP] = OpInfo{
+    op_info_vec[ADD_VV_OP as usize] = OpInfo{
         name           : "add_vv".to_string(),
         forward_0      : float_forward_0_add_vv,
         forward_1      : float_forward_1_add_vv,
